Gracefully degrade and warn if bad index
This should be an unlikely situation, but it seems to be happening. A keyword that is being used for search, is not found in the list of keyword indices. This could be because of stemming?

So the plugin continues working for user, this change will skip such indices. It will also print to the console as a warning to help with debugging why this is taking place.